Fixed script compilation

+++ src/main/kotlin/neo/CM/CollisionModel_contents.kt
@@ -20,11 +20,11 @@ object CollisionModel_contents {
      ================
      */
     fun CM_SetTrmEdgeSidedness(edge: cm_edge_s, bpl: idPluecker, epl: idPluecker, bitNum: Int) {
-        if (0 == edge.sideSet and (1 shl bitNum)) {
+        if (0L == edge.sideSet and (1L shl bitNum)) {
             val fl: Float
             fl = bpl.PermutedInnerProduct(epl)
-            edge.side = edge.side and (1 shl bitNum).inv() or (Math_h.FLOATSIGNBITSET(fl) shl bitNum)
-            edge.sideSet = edge.sideSet or (1 shl bitNum)
+            edge.side = edge.side and (1L shl bitNum).inv() or (Math_h.FLOATSIGNBITSET(fl).toLong() shl bitNum)
+            edge.sideSet = edge.sideSet or (1L shl bitNum)
         }
     }
 
@@ -34,15 +34,15 @@ object CollisionModel_contents {
      ================
      */
     fun CM_SetTrmPolygonSidedness(v: cm_vertex_s, plane: idPlane, bitNum: Int) {
-        if (0 == v.sideSet and (1 shl bitNum)) {
+        if (0L == v.sideSet and (1L shl bitNum)) {
             val fl: Float
             fl = plane.Distance(v.p)
             /* cannot use float sign bit because it is undetermined when fl == 0.0f */if (fl < 0.0f) {
-                v.side = v.side or (1 shl bitNum)
+                v.side = v.side or (1L shl bitNum)
             } else {
-                v.side = v.side and (1 shl bitNum).inv()
+                v.side = v.side and (1L shl bitNum).inv()
             }
-            v.sideSet = v.sideSet or (1 shl bitNum)
+            v.sideSet = v.sideSet or (1L shl bitNum)
         }
     }
 }
